move component did mount logic into component
